+++ src/components/CustomCodeEditor/CustomCodeEditor.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { StandardEditorProps } from '@grafana/data';
-import { CodeEditor, Monaco } from '@grafana/ui';
-import { CodeEditorHeight, CodeLanguage } from '../../constants';
+import { getTemplateSrv } from '@grafana/runtime';
+import { CodeEditor, CodeEditorSuggestionItem, CodeEditorSuggestionItemKind } from '@grafana/ui';
+import { CodeEditorHeight, CodeEditorSuggestions, CodeLanguage } from '../../constants';
 
 /**
  * Monaco
@@ -21,14 +22,39 @@ export const CustomCodeEditor: React.FC<Props> = ({ value, item, onChange }) => 
    * Language
    */
   const language = item.settings?.language || CodeLanguage.JAVASCRIPT;
+  const variables = getTemplateSrv().getVariables();
 
   /**
    * Format On Mount
    */
-  const onEditorMount = (editor: monacoType.editor.IStandaloneCodeEditor, monaco: Monaco) => {
+  const onEditorMount = (editor: monacoType.editor.IStandaloneCodeEditor) => {
     setTimeout(() => {
       editor.getAction('editor.action.formatDocument').run();
     }, 100);
+  };
+
+  /**
+   * Suggestions
+   *
+   * Will skip adding suggestions if settings are not defined to avoid duplicates
+   */
+  const getSuggestions = (): CodeEditorSuggestionItem[] => {
+    if (!item.settings?.suggestions) {
+      return [];
+    }
+
+    /**
+     * Add Variables
+     */
+    const suggestions = variables.map((variable) => {
+      return {
+        label: `\$\{${variable.name}\}`,
+        kind: CodeEditorSuggestionItemKind.Property,
+        detail: variable.description ? variable.description : variable.label,
+      };
+    });
+
+    return [...CodeEditorSuggestions, ...suggestions];
   };
 
   /**
@@ -39,14 +65,18 @@ export const CustomCodeEditor: React.FC<Props> = ({ value, item, onChange }) => 
       <CodeEditor
         language={language}
         showLineNumbers={true}
-        showMiniMap={true}
+        showMiniMap={(value && value.length) > 100}
         value={value}
         height={`${CodeEditorHeight}px`}
         onBlur={(code) => {
           onChange(code);
         }}
+        onSave={(code) => {
+          onChange(code);
+        }}
         monacoOptions={{ formatOnPaste: true, formatOnType: true }}
         onEditorDidMount={onEditorMount}
+        getSuggestions={getSuggestions}
       />
     </div>
   );